tests: ring_buffer: split suite by API variant
Replace the wildcard source glob with explicit sources so the suite can be
run both on the deprecated and non-deprecated APIs.
