+++ tests/lib/ringbuffer/CMakeLists.txt
@@ -4,5 +4,17 @@ cmake_minimum_required(VERSION 3.28.0)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(ringbuffer)
 
-FILE(GLOB app_sources src/*.c)
-target_sources(app PRIVATE ${app_sources})
+target_sources(app PRIVATE
+    src/main.c
+    src/concurrent.c
+  )
+
+target_sources_ifndef(CONFIG_RING_BUFFER
+    app PRIVATE
+    src/index_math.c
+  )
+
+target_sources_ifdef(CONFIG_RING_BUFFER
+    app PRIVATE
+    src/deprecated.c
+  )
