+++ tests/lib/ringbuffer/prj.conf
@@ -7,6 +7,3 @@ CONFIG_TEST_RANDOM_GENERATOR=y
 CONFIG_ENTROPY_GENERATOR=y
 CONFIG_XOSHIRO_RANDOM_GENERATOR=y
 CONFIG_MP_MAX_NUM_CPUS=1
-
-# deprecation of item api
-CONFIG_DEPRECATION_TEST=y
