+++ tests/lib/ringbuffer/tests.yaml
@@ -10,9 +10,19 @@ tests:
       - native_sim
       - native_sim/native/64
 
+  libraries.ring_buffer.deprecated:
+    extra_configs:
+      - CONFIG_RING_BUFFER=y
+      - CONFIG_DEPRECATION_TEST=y
+    integration_platforms:
+      - native_sim
+      - native_sim/native/64
+
   libraries.ring_buffer.concurrent:
     platform_allow: qemu_x86
     extra_configs:
+      - CONFIG_RING_BUFFER=y
+      - CONFIG_DEPRECATION_TEST=y
       - CONFIG_SYS_CLOCK_TICKS_PER_SEC=100000
     integration_platforms:
       - qemu_x86
